third_party/blink/web_tests/editing/shadow/breaking-editing-boundary-with-table-expected.txt

CONSOLE WARNING: line 17: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
On the second time selecting <span>, it should not become non-contenteditable. If it becomes non-contenteditable, the selection from <span> to <table> will break editing boundaries and contain "a". This test checks the <span> does not changed to non-contenteditable element.

PASS contains(selectedString, "a") is false
PASS successfullyParsed is true

TEST COMPLETE
